requirements.txt
pandas>=1.4.0
numpy>=1.21.0
scikit-learn>=1.0.2
xgboost>=1.5.0
matplotlib>=3.5.0
PyYAML>=6.0
joblib>=1.1.0
rdkit>=2022.3.1
openai>=1.0.0
openpyxl>=3.0.10
tqdm>=4.64.0
